Rename to MockIndexer
