+++ test/mockindexer/files/.gitignore
