+++ test/mockindexer/irc/client.go
