+++ test/mockindexer/irc/handlers.go
