+++ test/mockindexer/irc/server.go
